+++ AdapterUnitTests/ChartboostAdapterTests/AUTChartboostInterstitialAdTests.m
@@ -152,7 +152,6 @@ typedef void (^AUTChartboostSetUpCompletionBlock)(CHBStartError *);
 
 - (void)testPresentation {
   [self mockSuccessfulAppStart];
-  OCMStub([_interstitialAdMock isCached]).andReturn(YES);
   UIViewController *controller = [[UIViewController alloc] init];
   OCMStub([_interstitialAdMock showFromViewController:controller])
       .andDo(^(NSInvocation *invocation) {
@@ -174,7 +173,6 @@ typedef void (^AUTChartboostSetUpCompletionBlock)(CHBStartError *);
 
 - (void)testDismiss {
   [self mockSuccessfulAppStart];
-  OCMStub([_interstitialAdMock isCached]).andReturn(YES);
 
   UIViewController *controller = [[UIViewController alloc] init];
   OCMStub([_interstitialAdMock showFromViewController:controller])
@@ -192,7 +190,6 @@ typedef void (^AUTChartboostSetUpCompletionBlock)(CHBStartError *);
 
 - (void)testClick {
   [self mockSuccessfulAppStart];
-  OCMStub([_interstitialAdMock isCached]).andReturn(YES);
 
   UIViewController *controller = [[UIViewController alloc] init];
   OCMStub([_interstitialAdMock showFromViewController:controller])
@@ -209,7 +206,6 @@ typedef void (^AUTChartboostSetUpCompletionBlock)(CHBStartError *);
 
 - (void)testClickWithError {
   [self mockSuccessfulAppStart];
-  OCMStub([_interstitialAdMock isCached]).andReturn(YES);
 
   UIViewController *controller = [[UIViewController alloc] init];
   OCMStub([_interstitialAdMock showFromViewController:controller])
@@ -227,7 +223,6 @@ typedef void (^AUTChartboostSetUpCompletionBlock)(CHBStartError *);
 
 - (void)testPresentationError {
   [self mockSuccessfulAppStart];
-  OCMStub([_interstitialAdMock isCached]).andReturn(YES);
   AUTKMediationInterstitialAdEventDelegate *eventDelegate =
       [self loadAdWithLocation:@"ad_location"];
   UIViewController *controller = [[UIViewController alloc] init];
@@ -247,21 +242,6 @@ typedef void (^AUTChartboostSetUpCompletionBlock)(CHBStartError *);
   XCTAssertEqual(eventDelegate.reportImpressionInvokeCount, 0);
 }
 
-- (void)testNotCachedError {
-  [self mockSuccessfulAppStart];
-  AUTKMediationInterstitialAdEventDelegate *eventDelegate =
-      [self loadAdWithLocation:@"ad_location"];
-  UIViewController *controller = [[UIViewController alloc] init];
-  [_interstitialDelegate presentFromViewController:controller];
-
-  XCTAssertNotNil(eventDelegate.didFailToPresentError);
-  XCTAssertEqual(eventDelegate.didFailToPresentError.code, GADMAdapterChartboostErrorAdNotCached);
-
-  // Must not trigger any presentation callbacks.
-  XCTAssertEqual(eventDelegate.willPresentFullScreenViewInvokeCount, 0);
-  XCTAssertEqual(eventDelegate.reportImpressionInvokeCount, 0);
-}
-
 - (void)testLoadAdFailsWhenStartFails {
   [self mockFailedAppStart];
 
@@ -275,7 +255,10 @@ typedef void (^AUTChartboostSetUpCompletionBlock)(CHBStartError *);
       [[AUTKMediationInterstitialAdConfiguration alloc] init];
   configuration.credentials = credentials;
 
-  NSError *expectedError = [[NSError alloc] initWithDomain:@"test_domain" code:1 userInfo:nil];
+  NSError *expectedError =
+      [[NSError alloc] initWithDomain:[GADMAdapterChartboostConstants errorDomain]
+                                 code:GADMAdapterChartboostErrorInitializationFailure
+                             userInfo:nil];
   AUTKWaitAndAssertLoadInterstitialAdFailure(_adapter, configuration, expectedError);
 }